+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Upgraded to create-inventory plugin v2.0.0.
 * Prevent check out when item with intellectual item status is scanned. Refs UICHKOUT-669.
 * Update to stripes v6. Refs UICHKOUT-687.
+* Proxy not allowed to checkout items for sponsor if proxy has patron block. Refs UICHKOUT-673.
 
 ## [5.0.0](https://github.com/folio-org/ui-checkout/tree/v5.0.0) (2020-10-12)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v4.0.1...v5.0.0)